refactor: restructure workspace to monorepo and API to clean architecture
