+++ services/api/cmd/api/main_test.go
@@ -1,0 +1,8 @@
+package main
+
+import "testing"
+
+func TestAPIEnvironment(t *testing.T) {
+	// Placeholder test to verify that the testing environment for the API cmd is set up correctly.
+	t.Log("API cmd environment test passed.")
+}
