+++ services/api/internal/handler/handler_test.go
@@ -1,0 +1,8 @@
+package handler
+
+import "testing"
+
+func TestHandlerLayerSetup(t *testing.T) {
+	// Placeholder test to illustrate where handler/HTTP tests should reside.
+	t.Log("Handler layer test passed.")
+}
