+++ services/api/internal/repository/models.go
@@ -2,7 +2,7 @@
 // versions:
 //   sqlc v1.30.0
 
-package db
+package repository
 
 import (
 	"database/sql"
