+++ services/api/internal/repository/occurrences.sql.go
@@ -3,7 +3,7 @@
 //   sqlc v1.30.0
 // source: occurrences.sql
 
-package db
+package repository
 
 import (
 	"context"
