+++ services/api/internal/service/service_test.go
@@ -1,0 +1,8 @@
+package service
+
+import "testing"
+
+func TestServiceLayerSetup(t *testing.T) {
+	// Placeholder test to illustrate where service/domain tests should reside.
+	t.Log("Service layer test passed.")
+}
