+++ services/api/sqlc.yaml
@@ -1,11 +1,11 @@
 version: "2"
 sql:
-  - schema: "internal/db/migration/000001_init_schema.up.sql"
-    queries: "internal/db/query/"
+  - schema: "sqlc/migration/000001_init_schema.up.sql"
+    queries: "sqlc/query/"
     engine: "postgresql"
     gen:
       go:
-        package: "db"
-        out: "internal/db/sqlc"
+        package: "repository"
+        out: "internal/repository"
         emit_json_tags: true
         emit_interface: true
