+++ services/api/sqlc/query/occurrences.sql
